Link more f_op TUs

+++ include/d/actor/d_a_player.h
@@ -267,7 +267,7 @@ public:
     int getParam0() const { return mParam0; }
     void setParam0(int value) { mParam0 = value; }
     void setParam1(int value) { mParam1 = value; }
-    void setParam2(int value) { mParam2 = value; }
+    void setParam2(s16 value) { mParam2 = value; }
     void setPos0(const cXyz* pos) { mDemoPos0 = *pos; }
     void setToolDemoType() { setDemoType(DEMO_TYPE_TOOL_e); }
     s16 getTimer() const { return mTimer; }
